<Refactor> layout.tsx 파일 이동
